requirements.txt
VTK==5.8.0
ansicolors==1.0.2
mayavi==4.3.0
nibabel==1.3.0
numpy==1.7.1
pynifti==0.20100607.1
scipy==0.12.0
PySide==1.2.1
